Run ucf_get_gp for all scales

+++ ucf_sports/dim_14/grass_points/grass-def.hpp
@@ -4,7 +4,7 @@ public:
     inline grass_points(const std::string in_path,
 			const std::string in_path_dataset,
 			const std::string in_actionNames,  
-			const int in_scale_factor, 
+			const float in_scale_factor, 
 			const int in_shift,
 			const int in_dim,		    
 			const int in_p
@@ -18,7 +18,7 @@ const std::string path;
 const std::string path_dataset;
 const std::string actionNames;
 
-const int scale_factor;
+const float scale_factor;
 const int shift;
 const int p;
 const int dim;
